+++ src/components/pages/main_page/MainPage.svelte
@@ -5,27 +5,20 @@
   import { delay } from '@tools/delay';
   import { writable, type Unsubscriber } from 'svelte/store';
   import { LANG_LIST, SCRIPT_LIST } from '@tools/lang_list';
-  import { load_parivartak_lang_data, lipi_parivartak_async } from '@tools/converter';
+  import LipiLekhikA, { load_parivartak_lang_data, lipi_parivartak_async } from '@tools/converter';
   import { LanguageIcon } from '@components/icons';
-  import User from './user/User.svelte';
   import { ensure_auth_access_status, get_id_token_info } from '@tools/auth_tools';
   import { browser } from '$app/environment';
-  import DisplayArea from './display/DisplayArea.svelte';
-  import { client_raw } from '@api/client';
+  import SargaDisplay from './display/SargaDisplay.svelte';
   import { get_possibily_not_undefined } from '@tools/kry';
-  import { BiEdit } from 'svelte-icons-pack/bi';
+  import { BiEdit, BiHelpCircle } from 'svelte-icons-pack/bi';
   import { scale, slide } from 'svelte/transition';
   import { TiArrowBackOutline, TiArrowForwardOutline } from 'svelte-icons-pack/ti';
   import { user_info } from '@state/main_page/user';
   import { goto } from '$app/navigation';
   import Select from '@components/Select.svelte';
   import { z } from 'zod';
-  import { createMutation, useQueryClient } from '@tanstack/svelte-query';
-  import { BsThreeDots } from 'svelte-icons-pack/bs';
-  import { popup } from '@skeletonlabs/skeleton';
-  import { RiDocumentFileExcel2Line } from 'svelte-icons-pack/ri';
-  import { download_file_in_browser } from '@tools/download_file_browser';
-  import { transliterate_xlxs_file } from '@tools/excel/transliterate_xlsx_file';
+  import { createMutation, createQuery, useQueryClient } from '@tanstack/svelte-query';
   import {
     kANDa_selected,
     sarga_selected,
@@ -33,10 +26,15 @@
     BASE_SCRIPT,
     viewing_script,
     trans_lang,
-    view_translation_status
+    view_translation_status,
+    edit_language_typer_status,
+    sanskrit_mode,
+    typing_assistance_modal_opened
   } from '@state/main_page/main_state';
-  import { load_english_translation, sarga_data } from '@state/main_page/data';
   import { user_allowed_langs } from '@state/main_page/user';
+  import { VscAccount } from 'svelte-icons-pack/vsc';
+  import { popup, SlideToggle } from '@skeletonlabs/skeleton';
+  import { BsKeyboard } from 'svelte-icons-pack/bs';
 
   let mounted = false;
 
@@ -173,64 +171,24 @@
     unsubscribers.forEach((unsub) => unsub());
   });
 
-  const download_excel_file = createMutation({
-    mutationKey: ['sarga', 'download_excel_data'],
-    mutationFn: async () => {
-      if (!browser) return;
-      // the method used below creates a url for both dev and prod
-      const ExcelJS = (await import('exceljs')).default;
-      const url = new URL('/data/ramayan/template/excel_file_template.xlsx', import.meta.url).href;
-      const req = await fetch(url);
-      const file_blob = await req.blob();
-      const workbook = new ExcelJS.Workbook();
-      await workbook.xlsx.load(await file_blob.arrayBuffer());
-      const worksheet = workbook.getWorksheet(1)!;
-      const COLUMN_FOR_DEV = 2;
-      const TEXT_START_ROW = 2;
-      const translation_texts: Map<string, Map<number, string>> = new Map();
-      //load english translations
-      const english_translation = await load_english_translation($kANDa_selected, $sarga_selected);
-      translation_texts.set('English', english_translation);
-      const shloka_count =
-        rAmAyaNa_map[$kANDa_selected - 1].sarga_data[$sarga_selected - 1].shloka_count;
-      // loading other language translations
-      const other_translations =
-        await client_raw.translations.get_all_langs_translations_per_sarga.query({
-          kANDa_num: $kANDa_selected,
-          sarga_num: $sarga_selected
-        });
-      for (let data of other_translations) {
-        if (!translation_texts.has(data.lang)) translation_texts.set(data.lang, new Map());
-        translation_texts.get(data.lang)!.set(data.shloka_num, data.text);
-      }
-
-      for (let i = 0; i < $sarga_data.data!.length; i++) {
-        worksheet.getCell(i + COLUMN_FOR_DEV, TEXT_START_ROW).value = $sarga_data.data![i];
-      }
-      await transliterate_xlxs_file(
-        workbook,
-        'all',
-        1,
-        COLUMN_FOR_DEV,
-        TEXT_START_ROW,
-        BASE_SCRIPT,
-        null,
-        translation_texts,
-        shloka_count
-      );
-
-      // saving file to output path
-      let sarga_name =
-        rAmAyaNa_map[$kANDa_selected - 1].sarga_data[$sarga_selected - 1].name_normal.split(
-          '\n'
-        )[0];
-      const buffer = await workbook.xlsx.writeBuffer();
-      const blob = new Blob([buffer], {
-        type: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
-      });
-      const downloadLink = URL.createObjectURL(blob);
-      download_file_in_browser(downloadLink, `${$sarga_selected}. ${sarga_name}.xlsx`);
-    }
+  // Language Typing for Schwa Deletion
+  $: sanskrit_mode_texts = createQuery({
+    queryKey: ['sanskrit_mode_texts'],
+    enabled: browser && $editing_status_on && $trans_lang !== '--',
+    queryFn: () =>
+      Promise.all(
+        ['राम्', 'राम'].map((text) => lipi_parivartak_async(text, BASE_SCRIPT, $trans_lang))
+      ),
+    placeholderData: ['राम्', 'राम']
+  });
+  onMount(() => {
+    unsubscribers.push(
+      sanskrit_mode_texts.subscribe(({ isFetching, isSuccess }) => {
+        if (!$editing_status_on || isFetching || !isSuccess) return;
+        const lng = LipiLekhikA.k.normalize($trans_lang);
+        $sanskrit_mode = (LipiLekhikA.k.akSharAH as any)[lng].sa;
+      })
+    );
   });
 </script>
 
@@ -249,7 +207,23 @@
         {/each}
       </select>
     </label>
-    <User />
+    <!-- User -->
+    <button
+      class="btn m-2 p-0"
+      use:popup={{
+        event: 'click',
+        target: 'user_info',
+        placement: 'left-start'
+      }}
+    >
+      <Icon class="hover:text-gray-6200 text-3xl dark:hover:text-gray-400" src={VscAccount} />
+    </button>
+    <div class="card z-40 px-1 py-2 shadow-2xl" data-popup="user_info">
+      {#await import('./user/User.svelte') then User}
+        <User.default />
+      {/await}
+    </div>
+    <!-- /User -->
   </div>
   <!-- svelte-ignore a11y-label-has-associated-control -->
   <label class="space-x-4">
@@ -286,36 +260,9 @@
         />
       </label>
       {#if $kANDa_selected !== 0 && $sarga_selected !== 0}
-        <button
-          title="Extra Options"
-          transition:scale
-          use:popup={{
-            event: 'click',
-            target: 'sarga_options',
-            placement: 'bottom'
-          }}
-          class="btn m-0 rounded-full p-[0.05rem] ring-2 ring-zinc-400 dark:ring-zinc-300"
-        >
-          <Icon class="text-2xl" src={BsThreeDots} />
-        </button>
-        <div class="card z-50 space-y-1 rounded-lg px-1 py-1 shadow-xl" data-popup="sarga_options">
-          <button
-            on:click={() => $download_excel_file.mutate()}
-            class="btn block w-full rounded-md px-2 py-1 hover:bg-gray-200 dark:hover:bg-gray-700"
-          >
-            <Icon
-              class="-mt-1 mr-1 text-2xl text-green-600 dark:text-green-400"
-              src={RiDocumentFileExcel2Line}
-            />
-            Download Excel File
-          </button>
-          <!-- <button
-            class="btn block w-full rounded-md px-2 py-1 text-start hover:bg-gray-200 dark:hover:bg-gray-700"
-          >
-            <Icon src={TrOutlineFileTypeTxt} class=" mr-1 text-2xl" />
-            Download Text File
-          </button> -->
-        </div>
+        {#await import('./display/SargaUtility.svelte') then SargaUtility}
+          <SargaUtility.default />
+        {/await}
       {/if}
     </div>
   {/if}
@@ -382,7 +329,37 @@
         {/if}
       {/if}
     </div>
-    <DisplayArea />
+    <div class="flex space-x-4">
+      {#if $editing_status_on}
+        <SlideToggle
+          name="edit_lang"
+          active="bg-primary-500"
+          class="hover:text-gray-500 dark:hover:text-gray-400"
+          bind:checked={$edit_language_typer_status}
+          size="sm"
+        >
+          <Icon src={BsKeyboard} class="text-4xl" />
+        </SlideToggle>
+        {#if $edit_language_typer_status && $sanskrit_mode_texts.isSuccess && !$sanskrit_mode_texts.isFetching}
+          <select
+            transition:scale
+            bind:value={$sanskrit_mode}
+            class="select m-0 w-28 text-clip px-1 py-1 text-sm"
+          >
+            <option value={1}>rAm ➔ {$sanskrit_mode_texts.data[0]}</option>
+            <option value={0}>rAm ➔ {$sanskrit_mode_texts.data[1]}</option>
+          </select>
+        {/if}
+        <button
+          class="btn rounded-md p-0 text-sm"
+          title={'Language Typing Assistance'}
+          on:click={() => ($typing_assistance_modal_opened = true)}
+        >
+          <Icon src={BiHelpCircle} class="mt-1 text-3xl text-sky-500 dark:text-sky-400" />
+        </button>
+      {/if}
+    </div>
+    <SargaDisplay />
   {/if}
 </div>
 <slot />
